feat: #147 improve backend structure
