+++ kodemy-backend/src/main/java/pl/sknikod/kodemybackend/infrastructure/category/CategoryService.java
@@ -5,9 +5,9 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import pl.sknikod.kodemybackend.exception.structure.ServerProcessingException;
 import pl.sknikod.kodemybackend.infrastructure.category.rest.SingleCategoryResponse;
-import pl.sknikod.kodemybackend.infrastructure.common.EntityDao;
 import pl.sknikod.kodemybackend.infrastructure.common.entity.Category;
 import pl.sknikod.kodemybackend.infrastructure.common.mapper.CategoryMapper;
+import pl.sknikod.kodemybackend.infrastructure.common.repository.CategoryRepository;
 
 import javax.transaction.Transactional;
 
@@ -15,11 +15,11 @@ import javax.transaction.Transactional;
 @AllArgsConstructor
 public class CategoryService {
     private final CategoryMapper categoryMapper;
-    private final EntityDao entityDao;
+    private final CategoryRepository categoryRepository;
 
     @Transactional
     public SingleCategoryResponse showCategoryInfo(Long categoryId) {
-        return Option.of(categoryMapper.map(entityDao.findCategoryById(categoryId)))
+        return Option.of(categoryMapper.map(categoryRepository.findCategoryById(categoryId)))
                 .getOrElseThrow(() -> new ServerProcessingException(ServerProcessingException.Format.PROCESS_FAILED, Category.class));
     }
 }
