+++ kodemy-backend/src/main/java/pl/sknikod/kodemybackend/infrastructure/common/mapper/GradeMapper.java
@@ -5,14 +5,13 @@ import pl.sknikod.kodemybackend.configuration.SecurityConfig;
 import pl.sknikod.kodemybackend.infrastructure.common.entity.Author;
 import pl.sknikod.kodemybackend.infrastructure.common.entity.Grade;
 import pl.sknikod.kodemybackend.infrastructure.common.entity.Material;
-import pl.sknikod.kodemybackend.infrastructure.material.rest.MaterialAddGradeRequest;
-import pl.sknikod.kodemybackend.infrastructure.material.rest.SingleGradeResponse;
+import pl.sknikod.kodemybackend.infrastructure.material.MaterialGradeUseCase;
 
 import java.util.Set;
 
 @Mapper(componentModel = "spring")
 public interface GradeMapper {
-    default Grade map(MaterialAddGradeRequest request, Material material, SecurityConfig.UserPrincipal author) {
+    default Grade map(MaterialGradeUseCase.MaterialAddGradeRequest request, Material material, SecurityConfig.UserPrincipal author) {
         var grade = new Grade();
         grade.setMaterial(material);
         grade.setAuthor(Author.map(author));
@@ -20,7 +19,7 @@ public interface GradeMapper {
         return grade;
     }
 
-    SingleGradeResponse map(Grade grade);
+    MaterialGradeUseCase.GradePageable map(Grade grade);
 
-    Set<SingleGradeResponse> map(Set<Grade> grades);
+    Set<MaterialGradeUseCase.GradePageable> map(Set<Grade> grades);
 }
