+++ kodemy-backend/src/main/java/pl/sknikod/kodemybackend/infrastructure/common/repository/MaterialRepository.java
@@ -1,11 +1,13 @@
 package pl.sknikod.kodemybackend.infrastructure.common.repository;
 
+import io.vavr.control.Option;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import pl.sknikod.kodemybackend.exception.structure.NotFoundException;
 import pl.sknikod.kodemybackend.infrastructure.common.entity.Material;
 import pl.sknikod.kodemybackend.infrastructure.material.rest.SearchFields;
 
@@ -23,4 +25,11 @@ public interface MaterialRepository extends JpaRepository<Material, Long> {
             "(:#{#searchFields.title} IS NULL OR m.title >= :#{#searchFields.title}) AND " +
             "(:#{#searchFields.createdBy} IS NULL OR m.author.username <= :#{#searchFields.createdBy})")
     Page<Material> searchMaterials(SearchFields searchFields, PageRequest pageRequest);
+
+    default Material findMaterialById(Long materialId) {
+        return Option.ofOptional(findById(materialId))
+                .getOrElseThrow(() ->
+                        new NotFoundException(NotFoundException.Format.ENTITY_ID, Material.class, materialId)
+                );
+    }
 }
