--- kodemy-backend/src/main/java/pl/sknikod/kodemybackend/infrastructure/material/rest/MaterialAddGradeRequest.java
@@ -1,14 +1,0 @@
-package pl.sknikod.kodemybackend.infrastructure.material.rest;
-
-import lombok.Data;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-
-@Data
-public class MaterialAddGradeRequest {
-    @NotNull
-    @Pattern(regexp = "^[0-5](\\.[05])?$")
-    private String grade;
-
-}
