+++ kodemy-backend/src/main/java/pl/sknikod/kodemybackend/infrastructure/material/rest/MaterialGradeController.java
@@ -8,27 +8,28 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.RestController;
-import pl.sknikod.kodemybackend.infrastructure.material.MaterialService;
+import pl.sknikod.kodemybackend.infrastructure.material.MaterialGradeUseCase;
 
 import java.util.Objects;
 
 @RestController
 @AllArgsConstructor
 public class MaterialGradeController implements MaterialGradeControllerDefinition {
-    private final MaterialService materialService;
+    private final MaterialGradeUseCase materialGradeUseCase;
 
     @Override
     @PreAuthorize("isAuthenticated()")
-    public void addGrade(MaterialAddGradeRequest body, Long materialId) {
-        materialService.addGrade(materialId, body);
+    public void addGrade(MaterialGradeUseCase.MaterialAddGradeRequest body, Long materialId) {
+        materialGradeUseCase.addGrade(materialId, body);
     }
 
     @Override
-    public ResponseEntity<Page<SingleGradeResponse>> showGrades(int size, int page, String sort, Sort.Direction sortDirection, GradeMaterialSearchFields searchFields) {
+    public ResponseEntity<Page<MaterialGradeUseCase.GradePageable>> showGrades(int size, int page, String sort, Sort.Direction sortDirection, GradeMaterialSearchFields searchFields) {
         var pageRequest = PageRequest.of(page, size, sortDirection, sort);
         var searchFieldsParam = Objects.isNull(searchFields) ? new GradeMaterialSearchFields() : searchFields;
-        return ResponseEntity.status(HttpStatus.OK).body(
-                materialService.showGrades(pageRequest, searchFieldsParam)
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(
+                materialGradeUseCase.showGrades(pageRequest, searchFieldsParam)
         );
     }
 }
